+++ src/__tests__/Model.js
@@ -3,7 +3,7 @@ import { toJS, observable } from 'mobx';
 import MockAdapter from 'axios-mock-adapter';
 import _ from 'lodash';
 import { Model, BinderApi, Casts } from '../';
-import { compareObjectsIgnoringNegativeIds } from "./helpers";
+import { compareObjectsIgnoringNegativeIds } from "./helpers/helpers";
 import {
     Animal,
     AnimalStore,
